+++ hermes_shm/test/unit/data_structures/containers/list.h
@@ -28,8 +28,8 @@ class ListTestSuite {
   : obj_(obj), alloc_(alloc) {}
 
   /// Emplace elements
-  void EmplaceTest(int count = 30) {
-    for (int i = 0; i < count; ++i) {
+  void EmplaceTest(size_t count = 30) {
+    for (size_t i = 0; i < count; ++i) {
       CREATE_SET_VAR_TO_INT_OR_STRING(T, var, i);
       obj_.emplace_back(var);
     }
@@ -37,8 +37,8 @@ class ListTestSuite {
   }
 
   /// Forward iterator
-  void ForwardIteratorTest(int count = 30) {
-    int fcur = 0;
+  void ForwardIteratorTest(size_t count = 30) {
+    size_t fcur = 0;
     for (auto num : obj_) {
       CREATE_SET_VAR_TO_INT_OR_STRING(T, fcur_conv, fcur);
       REQUIRE(*num == fcur_conv);
@@ -47,9 +47,9 @@ class ListTestSuite {
   }
 
   /// Constant Forward iterator
-  void ConstForwardIteratorTest(int count = 30) {
+  void ConstForwardIteratorTest(size_t count = 30) {
     const Container &obj = obj_;
-    int fcur = 0;
+    size_t fcur = 0;
     for (auto iter = obj.cbegin(); iter != obj.cend(); ++iter) {
       hipc::Ref<T> num = *iter;
       CREATE_SET_VAR_TO_INT_OR_STRING(T, fcur_conv, fcur);
@@ -60,14 +60,14 @@ class ListTestSuite {
 
   /// Copy constructor
   void CopyConstructorTest() {
-    int count = obj_.size();
+    size_t count = obj_.size();
     auto cpy = hipc::make_uptr<Container>(obj_);
     VerifyCopy(obj_, *cpy, count);
   }
 
   /// Copy assignment
   void CopyAssignmentTest() {
-    int count = obj_.size();
+    size_t count = obj_.size();
     auto cpy = hipc::make_uptr<Container>();
     *cpy = obj_;
     VerifyCopy(obj_, *cpy, count);
@@ -75,7 +75,7 @@ class ListTestSuite {
 
   /// Move constructor
   void MoveConstructorTest() {
-    int count = obj_.size();
+    size_t count = obj_.size();
     auto cpy = hipc::make_uptr<Container>(std::move(obj_));
     VerifyMove(obj_, *cpy, count);
     obj_ = std::move(*cpy);
@@ -84,7 +84,7 @@ class ListTestSuite {
 
   /// Move assignment
   void MoveAssignmentTest() {
-    int count = obj_.size();
+    size_t count = obj_.size();
     auto cpy = hipc::make_uptr<Container>();
     (*cpy) = std::move(obj_);
     VerifyMove(obj_, *cpy, count);
@@ -95,7 +95,7 @@ class ListTestSuite {
   /// Emplace and erase front
   void EmplaceFrontTest() {
     CREATE_SET_VAR_TO_INT_OR_STRING(T, i0, 100);
-    int old_size = obj_.size();
+    size_t old_size = obj_.size();
     obj_.emplace_front(i0);
     REQUIRE(*obj_.front() == i0);
     REQUIRE(obj_.size() == old_size + 1);
@@ -146,13 +146,13 @@ class ListTestSuite {
   /// Verify copy construct/assign worked
   void VerifyCopy(Container &obj,
                   Container &cpy,
-                  int count) {
+                  size_t count) {
     REQUIRE(obj_.size() == count);
     REQUIRE(cpy.size() == count);
 
     // Verify obj
     {
-      int fcur = 0;
+      size_t fcur = 0;
       for (auto num : obj_) {
         CREATE_SET_VAR_TO_INT_OR_STRING(T, fcur_conv, fcur);
         REQUIRE(*num == fcur_conv);
@@ -162,7 +162,7 @@ class ListTestSuite {
 
     // Verify copy
     {
-      int fcur = 0;
+      size_t fcur = 0;
       for (auto num : cpy) {
         CREATE_SET_VAR_TO_INT_OR_STRING(T, fcur_conv, fcur);
         REQUIRE(*num == fcur_conv);
@@ -174,10 +174,10 @@ class ListTestSuite {
   /// Verify move worked
   void VerifyMove(Container &orig_obj,
                   Container &new_obj,
-                  int count) {
+                  size_t count) {
     // Verify move into cpy worked
     {
-      int fcur = 0;
+      size_t fcur = 0;
       REQUIRE(orig_obj.IsNull());
       REQUIRE(new_obj.size() == count);
       for (auto num : new_obj) {
